requirements.txt
streamlit
python-dotenv
requests
fastapi
uvicorn[standard]
pydantic
PyYAML
python-multipart
langchain
langchain-groq
requests-cache 
retry-requests 
numpy 
pandas
openmeteo-requests
tavily-python
sqlalchemy

